Add a test case for attribute update
