+++ package.json
@@ -4,7 +4,7 @@
   "description": "A flexible wrapper for gl-vao and gl-buffer that you can use to set up renderable WebGL geometries from a variety of different formats.",
   "main": "index.js",
   "scripts": {
-    "start": "budo test.js --open -- -t glslify",
+    "start": "budo test/test.js --open -- -t glslify -t brfs",
     "test": "standard"
   },
   "standard": {
@@ -37,7 +37,9 @@
     "ndarray": "^1.0.15",
     "normals": "^0.1.0",
     "standard": "^5.4.1",
-    "unindex-mesh": "0.0.0"
+    "unindex-mesh": "0.0.0",
+    "parse-wavefront-obj": "^1.0.1",
+    "brfs": "^1.4.3"
   },
   "repository": {
     "type": "git",
