+++ test/basicCloth.obj
@@ -1,0 +1,213 @@
+# Blender v2.74 (sub 0) OBJ File: 'basicCloth.blend'
+# www.blender.org
+o Plane
+v -3.512419 2.008027 3.512419
+v 3.512419 2.008027 3.512419
+v -3.512419 2.008027 -3.512419
+v 3.512419 2.008027 -3.512419
+v -3.512419 2.008027 0.000000
+v 0.000000 2.008027 3.512419
+v 3.512419 2.008027 0.000000
+v 0.000000 2.008027 -3.512419
+v 0.000000 2.008027 0.000000
+v -3.512419 2.008027 1.756210
+v 1.756210 2.008027 3.512419
+v 3.512419 2.008027 -1.756210
+v -1.756210 2.008027 -3.512419
+v -3.512419 2.008027 -1.756210
+v -1.756210 2.008027 3.512419
+v 3.512419 2.008027 1.756210
+v 1.756210 2.008027 -3.512419
+v 0.000000 2.008027 -1.756210
+v 0.000000 2.008027 1.756210
+v -1.756210 2.008027 0.000000
+v 1.756210 2.008027 0.000000
+v 1.756210 2.008027 1.756210
+v -1.756210 2.008027 1.756210
+v -1.756210 2.008027 -1.756210
+v 1.756210 2.008027 -1.756210
+v -3.512419 2.008027 2.634315
+v 2.634315 2.008027 3.512419
+v 3.512419 2.008027 -2.634315
+v -2.634315 2.008027 -3.512419
+v -3.512419 2.008027 -0.878105
+v -0.878105 2.008027 3.512419
+v 3.512419 2.008027 0.878105
+v 0.878105 2.008027 -3.512419
+v 0.000000 2.008027 -2.634315
+v 0.000000 2.008027 0.878105
+v -2.634315 2.008027 0.000000
+v 0.878105 2.008027 0.000000
+v -3.512419 2.008027 0.878105
+v 0.878105 2.008027 3.512419
+v 3.512419 2.008027 -0.878105
+v -0.878105 2.008027 -3.512419
+v -3.512419 2.008027 -2.634315
+v -2.634315 2.008027 3.512419
+v 3.512419 2.008027 2.634315
+v 2.634315 2.008027 -3.512419
+v 0.000000 2.008027 -0.878105
+v 0.000000 2.008027 2.634315
+v -0.878105 2.008027 0.000000
+v 2.634315 2.008027 0.000000
+v 1.756210 2.008027 0.878105
+v 1.756210 2.008027 2.634315
+v 0.878105 2.008027 1.756210
+v 2.634315 2.008027 1.756210
+v -1.756210 2.008027 0.878105
+v -1.756210 2.008027 2.634315
+v -2.634315 2.008027 1.756210
+v -0.878105 2.008027 1.756210
+v -1.756210 2.008027 -2.634315
+v -1.756210 2.008027 -0.878105
+v -2.634315 2.008027 -1.756210
+v -0.878105 2.008027 -1.756210
+v 1.756210 2.008027 -2.634315
+v 1.756210 2.008027 -0.878105
+v 0.878105 2.008027 -1.756210
+v 2.634315 2.008027 -1.756210
+v 2.634315 2.008027 -0.878105
+v 0.878105 2.008027 -0.878105
+v 0.878105 2.008027 -2.634315
+v -0.878105 2.008027 -0.878105
+v -2.634315 2.008027 -0.878105
+v -2.634315 2.008027 -2.634315
+v -0.878105 2.008027 2.634315
+v -2.634315 2.008027 2.634315
+v -2.634315 2.008027 0.878105
+v 2.634315 2.008027 2.634315
+v 0.878105 2.008027 2.634315
+v 0.878105 2.008027 0.878105
+v 2.634315 2.008027 0.878105
+v -0.878105 2.008027 0.878105
+v -0.878105 2.008027 -2.634315
+v 2.634315 2.008027 -2.634315
+s 1
+f 28 4 45
+f 34 8 41
+f 35 9 48
+f 32 7 49
+f 50 21 37
+f 51 22 52
+f 44 16 53
+f 54 20 36
+f 55 23 56
+f 47 19 57
+f 58 13 29
+f 59 24 60
+f 46 18 61
+f 62 17 33
+f 63 25 64
+f 40 12 65
+f 66 65 25
+f 49 66 63
+f 7 40 66
+f 67 64 18
+f 37 67 46
+f 21 63 67
+f 68 33 8
+f 64 68 34
+f 25 62 68
+f 69 61 24
+f 48 69 59
+f 9 46 69
+f 70 60 14
+f 36 70 30
+f 20 59 70
+f 71 29 3
+f 60 71 42
+f 24 58 71
+f 72 57 23
+f 31 72 55
+f 6 47 72
+f 73 56 10
+f 43 73 26
+f 15 55 73
+f 74 36 5
+f 56 74 38
+f 23 54 74
+f 75 53 22
+f 27 75 51
+f 2 44 75
+f 76 52 19
+f 39 76 47
+f 11 51 76
+f 77 37 9
+f 52 77 35
+f 22 50 77
+f 78 49 21
+f 53 78 50
+f 16 32 78
+f 79 48 20
+f 57 79 54
+f 19 35 79
+f 80 41 13
+f 61 80 58
+f 18 34 80
+f 81 45 17
+f 65 81 62
+f 12 28 81
+f 81 28 45
+f 80 34 41
+f 79 35 48
+f 78 32 49
+f 77 50 37
+f 76 51 52
+f 75 44 53
+f 74 54 36
+f 73 55 56
+f 72 47 57
+f 71 58 29
+f 70 59 60
+f 69 46 61
+f 68 62 33
+f 67 63 64
+f 66 40 65
+f 63 66 25
+f 21 49 63
+f 49 7 66
+f 46 67 18
+f 9 37 46
+f 37 21 67
+f 34 68 8
+f 18 64 34
+f 64 25 68
+f 59 69 24
+f 20 48 59
+f 48 9 69
+f 30 70 14
+f 5 36 30
+f 36 20 70
+f 42 71 3
+f 14 60 42
+f 60 24 71
+f 55 72 23
+f 15 31 55
+f 31 6 72
+f 26 73 10
+f 1 43 26
+f 43 15 73
+f 38 74 5
+f 10 56 38
+f 56 23 74
+f 51 75 22
+f 11 27 51
+f 27 2 75
+f 47 76 19
+f 6 39 47
+f 39 11 76
+f 35 77 9
+f 19 52 35
+f 52 22 77
+f 50 78 21
+f 22 53 50
+f 53 16 78
+f 54 79 20
+f 23 57 54
+f 57 19 79
+f 58 80 13
+f 24 61 58
+f 61 18 80
+f 62 81 17
+f 25 65 62
+f 65 12 81
